Added the option to include a token in Misc->printTable
The form in printTable might need a token under some circumstances, but
the use in all_db.php doesn't.  I've added a list of files that use this
feature to the bottom of the todo list.  If I get through all of them,
and they don't need tokens, I'll remove the option in Misc->printTable.