Partial commit for query results work
This will probably NEVER go live -- it's clearly becoming way too
complicated to pull all this stuff out. The files it's meant to replace
(sql.php and display.php) are used all over the codebase, and a change
this big is something that definitely shouldn't be part of a security
fix.

Committing here mostly in case I want to look back at this work for some
other reason.